https://docs.microsoft.com/en-us/aspnet/core/blazor/components/?view=aspnetcore-6.0#overwritten-parameters
"Important

Our general guidance is not to create components that directly write to their own parameters after the component is rendered for the first time."

WTF Microsoft? So child component are suposed to be read only??

That is equivalent to making method call with class types force behave as struct copy them and never be able to see the change in the scope above.