i) Went back to having component set variables in one component only without child components. The overhead required for child components in terms of:   a) writing lines of code to establish the parameters   b) writing lines of code to establish eventcallback that triger parent to call render   not only remove the advantage of having code split into smaller more readable chunks but also create combinatoric explosion of ways how this establishing can go bad and introduce nasty bag in the app.
Making the use of child components inefective sadly, the disadvatanges are at least as problematic as the advantages.

Strange decision on Microsoft's side. Algorithm to generalise the code solution is obvious.

ii) Added conditional disabling of range and operator radio buttons so that user cannot create RangeOperator variable, which is undefined in terms of meaning.